fix: Keep Mozilla's plural rule number for zero in Baltic, Shuar and Welsh
BREAKING CHANGE: These three language families used their last rule for the cardinal zero. It was not consistent with the result of the original Mozilla's rule evaluation. I moved the zero rule to the beginning for these three families to avoid mistakes from this only difference.

If you write language packs for these language groups:

* Baltic (Latvian, Latgalian)
* Ecuador indigenous languages (Shuar)
* Welsh

Move the string for the plural form for zero to the beginning of the string array.

The Semitic (Arabic) plural rules still put the zero plural form as the last one as the Mozilla's original does.